Object Libraries, part 5 (#2744)
* Object libraries: config, crypto, all_filters, checksum_filters, compression_filter, encryption_filters, filter_pipeline, parse_argument, tile
Extracted some parsing functions from util.* and moved them to new sources parse_argument.*, breaking what would otherwise be excessive link dependencies. Adjusted some code to accommodate these changes, including some untyped configuration calls.

* clang-format

* Adjusted build for Win32/non-MSVC

* Removed excess qualification in tile.h

* More adjustments to #include changes. Moved endian, hash functions out of util.h

* clang-format

* Removed extraneous definition that wasn't deleted after moving its declaration.

* Removed extraneous #include.

* Remove incorrect `assert`

* Adjusted logging headers in a few places.

* Move inclusion of Azure header <blob/blob_client.h> from `azure.h` to `azure.cc`.

* clang-format

* Rearrange #include in azure.cc

* Mark recently-added test "Sparse global order reader: qc removes full tile" as `[!mayfail]` while the failure it exposes is being investigated.

* Remove [!mayfail] marking from previous commit.

* In the `crypto` object library, suppress warning for MD5, which was deprecated in OpenSSL 3.

* Object libraries: datum, dimension, misc_types
Unit test executables: unit_datum, unit_array_schema
New classes: `UntypedDatum`, `DynamicTypedDatum` to refer to existing untyped data. Defined a stream output operator for `DynamicTypedDatum`. Rewrote `Writer::coords_to_str()` with new functions. Removed `Dimension::coord_to_str()`, no longer called.
Removed dependency of `Dimension` on `ResultCoords` and `QueryBuffer` classes.
Moved two overloads for `Dimension::map_to_uint64` to standalone functions in namespace `tiledb::sm::hilbert_order`. The overloads had arguments `ResultCoords` and `QueryBuffer`.
Added QueryBuffer member functions: fixed_size_datum_at(), varying_size_datum_at(), dimension_datum_at().
Moved `common_prefix_size` from `utils.*` to `stdx_string.*`

* Change return type of `Dimension::coord_size()` to `size_t` and inline the function.

* Add top-level dependencies for new unit test runners.

* Made some variables consistent with the return type of `Dimension::coord_size()`

* Add header to ensure std::min is declared.

* Remove fail-fast for Windows CI.

* Add #include <string_view> one place where it's necessary for MSVC2017

+++ tiledb/sm/fragment/fragment_metadata.cc
@@ -1049,7 +1049,7 @@ uint64_t FragmentMetadata::footer_size_v3_v4() const {
 uint64_t FragmentMetadata::footer_size_v5_v6() const {
   auto dim_num = array_schema_->dim_num();
   auto num = array_schema_->attribute_num() + dim_num + 1;
-  uint64_t domain_size = 0;
+  size_t domain_size = 0;
 
   if (non_empty_domain_.empty()) {
     // For var-sized dimensions, this function would be called only upon
